+++ src/main/java/domain/participant/Players.java
@@ -1,7 +1,7 @@
 package domain.participant;
 
-import domain.Deck;
-import domain.GamePoint;
+import domain.game.Deck;
+import domain.game.GamePoint;
 
 import java.util.List;
 import java.util.stream.Collectors;
