+++ src/test/java/domain/DeckTest.java
@@ -1,6 +1,7 @@
 package domain;
 
 import domain.card.Card;
+import domain.game.Deck;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
