+++ src/test/java/domain/GamePointTest.java
@@ -3,6 +3,7 @@ package domain;
 import domain.card.Card;
 import domain.card.CardNumber;
 import domain.card.CardShape;
+import domain.game.GamePoint;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -22,11 +23,11 @@ public class GamePointTest {
         );
 
         final GamePoint gamePoint = new GamePoint(data);
-        assertThat(gamePoint.getPoint()).isEqualTo(9);
+        assertThat(gamePoint.getPoint()).isSameAs(9);
     }
 
     @Test
-    @DisplayName("A는 1혹은 11로 계산할 수 있으면 해당 카드들의 최선 값을 반환한다.")
+    @DisplayName("A는 1혹은 11로 계산할 수 있으면 해당 카드들의 최적값을 반환한다.")
     void optimizeCalculateTest1() {
         final List<Card> data = List.of(
                 Card.create(CardShape.CLOVER, CardNumber.of(1)),
@@ -35,11 +36,11 @@ public class GamePointTest {
         );
 
         final GamePoint gamePoint = new GamePoint(data);
-        assertThat(gamePoint.getPoint()).isEqualTo(16);
+        assertThat(gamePoint.getPoint()).isSameAs(16);
     }
 
     @Test
-    @DisplayName("카드 계산 값이 21이 넘었을 경우 burst처리 한다.")
+    @DisplayName("카드 계산 값이 21이 넘었을 경우 bust(0점)처리 한다.")
     void optimizeCalculateTest2() {
         final List<Card> data = List.of(
                 Card.create(CardShape.CLOVER, CardNumber.of(12)),
@@ -47,7 +48,7 @@ public class GamePointTest {
                 Card.create(CardShape.HEART, CardNumber.of(2))
         );
         final GamePoint gamePoint = new GamePoint(data);
-        assertThat(gamePoint.getPoint()).isEqualTo(0);
+        assertThat(gamePoint.getPoint()).isSameAs(0);
     }
 
     @Test
@@ -59,7 +60,7 @@ public class GamePointTest {
                 Card.create(CardShape.HEART, CardNumber.of(1))
         );
         final GamePoint gamePoint = new GamePoint(data);
-        assertThat(gamePoint.getPoint()).isEqualTo(21);
+        assertThat(gamePoint.getPoint()).isSameAs(21);
     }
 
     @Test
@@ -74,6 +75,6 @@ public class GamePointTest {
                 Card.create(CardShape.HEART, CardNumber.of(1))
         );
         final GamePoint gamePoint = new GamePoint(data);
-        assertThat(gamePoint.getPoint()).isEqualTo(16);
+        assertThat(gamePoint.getPoint()).isSameAs(16);
     }
 }
